+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `Function.afterConstituted(task)` and `Function.hasConstituted()` methods
 * Add JSON-Dry support for `Map` and `Set`
+* Add the `HashSet` class
 
 ## 0.7.19 (2022-06-19)